Remove Pipeline class and move its methods into ProcessGraph
