+++ src/relion/_parser/processgraph.py
@@ -1,4 +1,5 @@
 import collections.abc
+from graphviz import Digraph
 from relion._parser.processnode import ProcessNode
 
 
@@ -104,6 +105,27 @@ class ProcessGraph(collections.abc.Sequence):
         for r in for_removal:
             connected_graphs.remove(r)
         return connected_graphs
+
+    def _split_connected(self, connected_dict, origin, origins_dict):
+        connected_graphs = self.split_connected()
+        ancillary_count = 1
+        for cg in connected_graphs:
+            if origin in cg:
+                connected_dict["main"] = cg
+                origins_dict["main"] = self[self.index(origin)]
+            else:
+                connected_dict[f"ancillary:{ancillary_count}"] = cg
+                origins_dict[f"ancillary:{ancillary_count}"] = cg.find_origins()[0]
+                ancillary_count += 1
+
+    def show_all_nodes(self):
+        digraph = Digraph(format="svg")
+        digraph.attr(rankdir="LR")
+        for node in self:
+            digraph.node(name=str(node._path))
+            for next_node in node:
+                digraph.edge(str(node._path), str(next_node._path))
+        digraph.render("pipeline.gv")
 
     def wipe(self):
         self._node_list = []
